requirements.txt
streamlit
ollama
reportlab
pillow
PyYAML
python-dotenv